+++ br/pkg/metautil/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
         "//br/pkg/storage",
         "//br/pkg/summary",
         "//pkg/parser/model",
-        "//pkg/statistics/handle/storage",
+        "//pkg/statistics/handle/util",
         "//pkg/tablecodec",
         "//pkg/util/encrypt",
         "@com_github_docker_go_units//:go-units",
