+++ pkg/executor/BUILD.bazel
@@ -165,7 +165,6 @@ go_library(
         "//pkg/statistics",
         "//pkg/statistics/handle",
         "//pkg/statistics/handle/cache",
-        "//pkg/statistics/handle/storage",
         "//pkg/statistics/handle/util",
         "//pkg/store/driver/backoff",
         "//pkg/store/driver/txn",
@@ -409,8 +408,8 @@ go_test(
         "//pkg/sessiontxn",
         "//pkg/sessiontxn/staleread",
         "//pkg/statistics",
-        "//pkg/statistics/handle/globalstats",
         "//pkg/statistics/handle/storage",
+        "//pkg/statistics/handle/util",
         "//pkg/store/copr",
         "//pkg/store/driver/error",
         "//pkg/store/helper",
