+++ pkg/server/handler/optimizor/BUILD.bazel
@@ -19,7 +19,7 @@ go_library(
         "//pkg/session",
         "//pkg/sessionctx/variable",
         "//pkg/statistics/handle",
-        "//pkg/statistics/handle/storage",
+        "//pkg/statistics/handle/util",
         "//pkg/table",
         "//pkg/types",
         "//pkg/util",
@@ -56,8 +56,7 @@ go_test(
         "//pkg/server/internal/testutil",
         "//pkg/server/internal/util",
         "//pkg/session",
-        "//pkg/statistics/handle/globalstats",
-        "//pkg/statistics/handle/storage",
+        "//pkg/statistics/handle/util",
         "//pkg/store/mockstore/unistore",
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
