+++ pkg/statistics/handle/BUILD.bazel
@@ -31,7 +31,6 @@ go_library(
         "//pkg/statistics/handle/globalstats",
         "//pkg/statistics/handle/history",
         "//pkg/statistics/handle/lockstats",
-        "//pkg/statistics/handle/metrics",
         "//pkg/statistics/handle/storage",
         "//pkg/statistics/handle/usage",
         "//pkg/statistics/handle/util",
@@ -72,7 +71,6 @@ go_test(
         "//pkg/sessionctx/stmtctx",
         "//pkg/sessionctx/variable",
         "//pkg/statistics",
-        "//pkg/statistics/handle/globalstats",
         "//pkg/statistics/handle/internal",
         "//pkg/statistics/handle/storage",
         "//pkg/statistics/handle/util",
