+++ pkg/statistics/handle/history/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "//pkg/statistics/handle/storage",
         "//pkg/statistics/handle/util",
         "//pkg/util/logutil",
-        "//pkg/util/sqlexec",
         "@com_github_pingcap_errors//:errors",
         "@org_uber_go_zap//:zap",
     ],
