+++ pkg/statistics/handle/storage/BUILD.bazel
@@ -25,6 +25,7 @@ go_library(
         "//pkg/statistics",
         "//pkg/statistics/handle/cache",
         "//pkg/statistics/handle/lockstats",
+        "//pkg/statistics/handle/metrics",
         "//pkg/statistics/handle/util",
         "//pkg/types",
         "//pkg/util/chunk",
@@ -36,7 +37,6 @@ go_library(
         "@com_github_klauspost_compress//gzip",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_pingcap_tipb//go-tipb",
         "@com_github_tikv_client_go_v2//oracle",
         "@org_uber_go_zap//:zap",
     ],
