+++ pkg/statistics/handle/util/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//pkg/util/sqlexec/mock",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_tipb//go-tipb",
         "@com_github_tiancaiamao_gp//:gp",
         "@com_github_tikv_client_go_v2//oracle",
     ],
